Added mpreal.hpp, my modified mpreal.h header
Working on implementing a method to accurately compute the intersection order between a line and a list of quadrics; all done except for the cases requiring extra precision.

Working on implementing a polynomial class which will correctly evaluate roots and products and such
Added a compilation test for this.

Templated evaluatePoint to support arbitrary floating point types
Need to use the typecast header to perform operations in the form with better precision

Added another test for line intersection calculations; had to change minimum allowed precision